Move Java cv to its own JNI class, move raw to separate package
