+++ cscore/src/main/java/edu/wpi/cscore/CameraServerJNI.java
@@ -11,16 +11,13 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.function.Consumer;
 
-import org.opencv.core.Core;
-
+import edu.wpi.cscore.raw.RawFrame;
 import edu.wpi.first.wpiutil.RuntimeLoader;
 
 public class CameraServerJNI {
   static boolean libraryLoaded = false;
-  static boolean cvLibraryLoaded = false;
 
   static RuntimeLoader<CameraServerJNI> loader = null;
-  static RuntimeLoader<Core> cvLoader = null;
 
   static {
     if (!libraryLoaded) {
@@ -32,18 +29,6 @@ public class CameraServerJNI {
         System.exit(1);
       }
       libraryLoaded = true;
-    }
-
-    String opencvName = Core.NATIVE_LIBRARY_NAME;
-    if (!cvLibraryLoaded) {
-      try {
-        cvLoader = new RuntimeLoader<>(opencvName, RuntimeLoader.getDefaultExtractionRoot(), Core.class);
-        cvLoader.loadLibraryHashed();
-      } catch (IOException ex) {
-        ex.printStackTrace();
-        System.exit(1);
-      }
-      cvLibraryLoaded = true;
     }
   }
 
@@ -71,7 +56,6 @@ public class CameraServerJNI {
   public static native int createUsbCameraPath(String name, String path);
   public static native int createHttpCamera(String name, String url, int kind);
   public static native int createHttpCameraMulti(String name, String[] urls, int kind);
-  public static native int createCvSource(String name, int pixelFormat, int width, int height, int fps);
   public static native int createRawSource(String name, int pixelFormat, int width, int height, int fps);
 
   //
@@ -124,9 +108,8 @@ public class CameraServerJNI {
   public static native String[] getHttpCameraUrls(int source);
 
   //
-  // OpenCV Source Functions
+  // Image Source Functions
   //
-  public static native void putSourceFrame(int source, long imageNativeObj);
   public static native void putRawSourceFrameBB(int source, ByteBuffer data, int width, int height, int pixelFormat, int totalData);
   public static native void putRawSourceFrame(int source, long data, int width, int height, int pixelFormat, int totalData);
   public static void putRawSourceFrame(int source, RawFrame raw) {
@@ -142,10 +125,8 @@ public class CameraServerJNI {
   // Sink Creation Functions
   //
   public static native int createMjpegServer(String name, String listenAddress, int port);
-  public static native int createCvSink(String name);
+
   public static native int createRawSink(String name);
-  //public static native int createCvSinkCallback(String name,
-  //                            void (*processFrame)(long time));
 
   //
   // Sink Functions
@@ -170,20 +151,18 @@ public class CameraServerJNI {
   public static native int getMjpegServerPort(int sink);
 
   //
-  // OpenCV Sink Functions
+  // Image Sink Functions
   //
   public static native void setSinkDescription(int sink, String description);
-  public static native long grabSinkFrame(int sink, long imageNativeObj);
-  public static native long grabSinkFrameTimeout(int sink, long imageNativeObj, double timeout);
 
   private static native long grabRawSinkFrameImpl(int sink, RawFrame rawFrame, long rawFramePtr, ByteBuffer byteBuffer, int width, int height, int pixelFormat);
   private static native long grabRawSinkFrameTimeoutImpl(int sink, RawFrame rawFrame, long rawFramePtr, ByteBuffer byteBuffer, int width, int height, int pixelFormat, double timeout);
 
   public static long grabSinkFrame(int sink, RawFrame rawFrame) {
-    return grabRawSinkFrameImpl(sink, rawFrame, rawFrame.m_framePtr, rawFrame.getDataByteBuffer(), rawFrame.getWidth(), rawFrame.getHeight(), rawFrame.getPixelFormat());
+    return grabRawSinkFrameImpl(sink, rawFrame, rawFrame.getFramePtr(), rawFrame.getDataByteBuffer(), rawFrame.getWidth(), rawFrame.getHeight(), rawFrame.getPixelFormat());
   }
   public static long grabSinkFrameTimeout(int sink, RawFrame rawFrame, double timeout) {
-    return grabRawSinkFrameTimeoutImpl(sink, rawFrame, rawFrame.m_framePtr, rawFrame.getDataByteBuffer(), rawFrame.getWidth(), rawFrame.getHeight(), rawFrame.getPixelFormat(), timeout);
+    return grabRawSinkFrameTimeoutImpl(sink, rawFrame, rawFrame.getFramePtr(), rawFrame.getDataByteBuffer(), rawFrame.getWidth(), rawFrame.getHeight(), rawFrame.getPixelFormat(), timeout);
   }
   public static native String getSinkError(int sink);
   public static native void setSinkEnabled(int sink, boolean enabled);
