+++ cscore/src/main/java/edu/wpi/cscore/CvSink.java
@@ -12,7 +12,7 @@ import org.opencv.core.Mat;
 /**
  * A sink for user code to accept video frames as OpenCV images.
  */
-public class CvSink extends VideoSink {
+public class CvSink extends ImageSink {
   /**
    * Create a sink for accepting OpenCV images.
    * WaitForFrame() must be called on the created sink to get each new
@@ -21,7 +21,7 @@ public class CvSink extends VideoSink {
    * @param name Source name (arbitrary unique identifier)
    */
   public CvSink(String name) {
-    super(CameraServerJNI.createCvSink(name));
+    super(CameraServerCvJNI.createCvSink(name));
   }
 
   /// Create a sink for accepting OpenCV images in a separate thread.
@@ -36,15 +36,6 @@ public class CvSink extends VideoSink {
   //       std::function<void(uint64_t time)> processFrame) {
   //  super(CameraServerJNI.createCvSinkCallback(name, processFrame));
   //}
-
-  /**
-   * Set sink description.
-   *
-   * @param description Description
-   */
-  public void setDescription(String description) {
-    CameraServerJNI.setSinkDescription(m_handle, description);
-  }
 
   /**
    * Wait for the next frame and get the image.
@@ -67,7 +58,7 @@ public class CvSink extends VideoSink {
    *         message); the frame time is in 1 us increments.
    */
   public long grabFrame(Mat image, double timeout) {
-    return CameraServerJNI.grabSinkFrameTimeout(m_handle, image.nativeObj, timeout);
+    return CameraServerCvJNI.grabSinkFrameTimeout(m_handle, image.nativeObj, timeout);
   }
 
   /**
@@ -78,24 +69,6 @@ public class CvSink extends VideoSink {
    *         message); the frame time is in 1 us increments.
    */
   public long grabFrameNoTimeout(Mat image) {
-    return CameraServerJNI.grabSinkFrame(m_handle, image.nativeObj);
-  }
-
-  /**
-   * Get error string.  Call this if WaitForFrame() returns 0 to determine
-   * what the error is.
-   */
-  public String getError() {
-    return CameraServerJNI.getSinkError(m_handle);
-  }
-
-  /**
-   * Enable or disable getting new frames.
-   * Disabling will cause processFrame (for callback-based CvSinks) to not
-   * be called and WaitForFrame() to not return.  This can be used to save
-   * processor resources when frames are not needed.
-   */
-  public void setEnabled(boolean enabled) {
-    CameraServerJNI.setSinkEnabled(m_handle, enabled);
+    return CameraServerCvJNI.grabSinkFrame(m_handle, image.nativeObj);
   }
 }
